+++ src/dfvue/dfvmain.py
@@ -20,11 +20,19 @@ The following classes are provided:
 
 History
     * Written Jul 2023 by Matthias Cuntz (mc (at) macu (dot) de)
-    * Moved to customtkinter, Jun 2024, Matthias
+    * Use CustomTkinter, Jun 2024, Matthias Cuntz
+    * Use mix of grid and pack layout manager, Jun 2024, Matthias Cuntz
+    * Use CustomTkinter only if installed, Jun 2024, Matthias Cuntz
 
 """
 import tkinter as tk
-import customtkinter as ctk
+import tkinter.ttk as ttk
+try:
+    from customtkinter import CTkTabview as Frame
+    ihavectk = True
+except ModuleNotFoundError:
+    from tkinter.ttk import Frame
+    ihavectk = False
 from .dfvscatter import dfvScatter
 
 
@@ -35,7 +43,7 @@ __all__ = ['dfvMain']
 # Window with plot panels
 #
 
-class dfvMain(ctk.CTkTabview):
+class dfvMain(Frame):
     """
     Main dfvue notebook window with the plotting panels.
 
@@ -56,11 +64,18 @@ class dfvMain(ctk.CTkTabview):
         self.master = master      # master window, i.e. root
         self.top    = master.top  # top window
 
-        stab = 'Scatter/Line'
-        self.add(stab)
-        itab = self.tab(stab)
-        itab.name   = self.name
-        itab.master = self.master
-        itab.top    = self.top
-        self.tab_scatter = dfvScatter(itab)
-        self.tab_scatter.grid(sticky=tk.W + tk.E + tk.S + tk.N)
+        if ihavectk:
+            stab = 'Scatter/Line'
+            self.add(stab)
+            itab = self.tab(stab)
+            itab.name   = self.name
+            itab.master = self.master
+            itab.top    = self.top
+            self.tab_scatter = dfvScatter(itab)
+            self.tab_scatter.grid(sticky=tk.W + tk.E + tk.S + tk.N)
+        else:
+            # Notebook for tabs for future plot types
+            self.tabs = ttk.Notebook(self)
+            self.tabs.grid(sticky=tk.W + tk.E + tk.S + tk.N)
+            self.tab_scatter = dfvScatter(self)
+            self.tabs.add(self.tab_scatter, text=self.tab_scatter.name)